Enhanced FreeCAD Coin selection mechanism
* SoFCSelection is modified to use SoFCUnifiedSelection for picking by
default. It now only does highlight and selection rendering.

* SoFCUnifiedSelection is extended to support Coin path based
selection/highlight instead of node based, so that it is possible to
distinguish the same Coin node in different paths.

* SoFCSelectionRoot is modified to assist path based picking, by
keeping a stak of currently travsed SoFCelectionRoot, and use that
stack as key for other SoNode to query selection context.

* SoBrepXXXSet is modified to make use of the above path based
selection context to draw selection and highlight.