PartGui: support partial rendering
Prepare for Part::LinkSub which can link to subelelement of a
Part::Feature and only render the linked subelements.

The implementation introduced a secondary selection context which
stores in the last encounted SoFCSelectionRoot node along the Coin node
path. The context is keyed using the first SoFCSelectionRoot that
requested partial linking, such that any further link to this partial
link will inherit the partially rendered appearance.