ViewProviderLink: add support for child view provider
A new property 'ChildViewProvider' allows user to embed a child view
provider.

Because ViewProviderLink is designed to work with any object as long as
it is installed with 'App::LinkBaseExtension'. Say, the object is
derived from 'Part::Feature'. So a child view provider can be added to
display the native shape, through a new display mode,  while still keep
the view provider linking capability.