PropertyView: add context menu action to print property
This action is intended to make it easy to find out how to access some
property in the property view through Python. It can be a little tricky
because the property name displayed in the view may not be the actual
property name. This action may also be helpful in case of
multi-selection, where the properties for each selected object will be
printed.