Gui: improve SoBrepFaceSet picking performance
Two Preference/View parameters controls the optimization,

SelectionPickThreshold, if zero disables the optimization. If non zero,
then if the number of parts is less than this threshold, calculate the
the bounding box per part, and linearly search for intersection per
part. Once found, perform picking of the faces with the part. If the
number of parts exceeds this threshold, then a boost::geometry::rtree
will be build to index the bounding box per part to accelerate the
search.

SelectionPickThreshold2 controls how to pick the faces within a part.
If zero, or the number of faces is smaller than this threshold, then
perform linear search of faces within a picked part. If greater, then
build an rtree to index the bounding box per face of a part to
accelerate the search.