Gui: add view command MapChildrenPlacement
This command enables mapping claimed children into parent's coordinate
space for all object. Link will show its children mapped regardless of
this option.

For non-Link objects, this mapping is implemented using a group coin3d
node in the view provider that contains all children root nodes
regardless of whether it implements getChildRoot() or not. This
internal group node is added to the mode switch node as its last child.
It will not be visible in normal situation, because there is no
corresponding display mode for it. MapChildrenPlacement will show it
using on top display, which renders the object using coin path,
overriding the active mode switch.

This patch also introduce a new property 'SyncGroupVisibility' to Link,
which allows it to show group children without override their
visibility.