Gui: track claimed children in ViewProviderDocumentObject
The claimed children tracking logic was originally implemented in tree
view. This patch moved implementation to ViewProviderDocumentObject
to makes the parent and child information more accessible to others,
similar to In/OutLsit in App namespace.

Python code can obtain the information with two new attributes in
ViewProviderDocumentObjectPy, ClaimedChildren and ClaimedBy