Spreadsheet: support cell binding
Cell binding allows one to bind a range of cells of one sheet to another
range of cells of an arbitary sheet, including any empty cells in the
range.

The binding is implemented with PropertyExpressionEngine, which binds a
special path '.cells.<start>.<end>' (see PropertySheet::setPathValue())
to a tuple(sheet, start, end). The start and end cell address of the
binding path must be constant, but can start and end in the tuple can
be other expressions.

A convenience menu action called 'Bind...' is added to SheetTableView
for easy binding.