Gui: fix tree view clear selection handling
Because current selection change trigged from 3D view is handled by a
timer. It is possible for the clear selection event to be handled
together with a subsequent element selection event. In case the same
object is selected previously with a  different element, the tree view
will contain the wrong element selection cache.

The fix is to clear tree view element selection cache immediately on
clear selection event.